Merge Test and TestResult models (#10)
Removed the Test model and combined it with TestResult. Also moved some properties to TestRun instead of duplicating them across all TestResult objects.

Add two-way reference between TestResult and TestRun to allow us easier queries from the front end. TestResult to TestRun is one-to-one, TestRun to TestResult is one-to-many. One potential con of this is if we needed to change that reference it's annoying/error prone to update it in both places (ex: moving a TestResult to another TestRun), but I don't believe we'll ever have to do that, so it should be all pros for us.