Add RunList and RunDetailView (#13)
At this point in time, this give us basic views for Runs and Run Details.
Still outstanding:

The results data is incomplete. Need to determine how to get the result data when required.
All test views are broken due to route changes. Will update in a different PR.
Deeplinking to a detail view page is currently broken